jet-audio-monitor/python/requirements.txt
grpcio==1.35.0
grpcio-tools==1.35.0
numpy==1.20.0
protobuf==3.14.0
six==1.15.0
